check phpmyadmin dir

+++ initial.wp-setup-mod_php.sh
@@ -201,12 +201,14 @@ fi
 
 # install phpMyAdmin
 cd /usr/share/
-/usr/bin/wget https://files.phpmyadmin.net/phpMyAdmin/${PHP_MY_ADMIN_VER}/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
-if [ -f phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip ]; then
-  /usr/bin/unzip /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
-  /bin/rm /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
-  /bin/rm /usr/share/phpMyAdmin
-  /bin/ln -s /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages /usr/share/phpMyAdmin
+if [ ! -d /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages ] ; then
+  /usr/bin/wget https://files.phpmyadmin.net/phpMyAdmin/${PHP_MY_ADMIN_VER}/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
+  if [ -f phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip ]; then
+    /usr/bin/unzip /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
+    /bin/rm /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages.zip
+    /bin/rm /usr/share/phpMyAdmin
+    /bin/ln -s /usr/share/phpMyAdmin-${PHP_MY_ADMIN_VER}-all-languages /usr/share/phpMyAdmin
+  fi
 fi
 
 #install DSaaS Client
